Prepare abstraction for handling relay candidates
