--- lib/ex_ice/if_discovery/inet.ex
@@ -1,8 +1,0 @@
-defmodule ExICE.IfDiscovery.Inet do
-  @moduledoc false
-
-  @behaviour ExICE.IfDiscovery
-
-  @impl true
-  defdelegate getifaddrs(), to: :inet
-end
