+++ lib/ex_ice/priv/if_discovery.ex
@@ -1,0 +1,6 @@
+defmodule ExICE.Priv.IfDiscovery do
+  @moduledoc false
+
+  # ifopts should be of type :inet.getifaddrs_ifopts() but this is an internal :inet type
+  @callback getifaddrs() :: {:ok, [{ifname :: charlist(), ifopts :: list()}]} | {:error, term()}
+end
