+++ lib/ex_ice/priv/if_discovery/inet.ex
@@ -1,0 +1,8 @@
+defmodule ExICE.Priv.IfDiscovery.Inet do
+  @moduledoc false
+
+  @behaviour ExICE.Priv.IfDiscovery
+
+  @impl true
+  defdelegate getifaddrs(), to: :inet
+end
