+++ test/candidate_test.exs
@@ -7,14 +7,24 @@ defmodule ExICE.CandidateTest do
     # FIXME socket shouldn't be nil
     ip = {192, 168, 1, 1}
     port = 12_345
-    %Candidate{foundation: f1} = Candidate.new(:host, ip, port, ip, port, nil)
-    %Candidate{foundation: f2} = Candidate.new(:host, ip, port, ip, port, nil)
+
+    %Candidate{foundation: f1} =
+      Candidate.new(:host, address: ip, port: port, base_address: ip, base_port: port)
+
+    %Candidate{foundation: f2} =
+      Candidate.new(:host, address: ip, port: port, base_address: ip, base_port: port)
+
     assert f1 == f2
 
     ip2 = {192, 168, 1, 2}
     port2 = 23_456
-    %Candidate{foundation: f1} = Candidate.new(:host, ip, port, ip, port, nil)
-    %Candidate{foundation: f2} = Candidate.new(:host, ip2, port2, ip2, port2, nil)
+
+    %Candidate{foundation: f1} =
+      Candidate.new(:host, address: ip, port: port, base_address: ip, base_port: port)
+
+    %Candidate{foundation: f2} =
+      Candidate.new(:host, address: ip2, port: port2, base_address: ip2, base_port: port2)
+
     assert f1 != f2
   end
 
@@ -23,7 +33,16 @@ defmodule ExICE.CandidateTest do
     port = 12_345
     expected_m_c = "936255739 1 UDP 1234 192.168.1.1 12345 typ host"
 
-    c = Candidate.new(:host, ip, port, ip, port, nil, priority: 1234)
+    c =
+      Candidate.new(:host,
+        address: ip,
+        port: port,
+        base_address: ip,
+        base_port: port,
+        priority: 1234,
+        transport_module: nil
+      )
+
     m_c = Candidate.marshal(c)
 
     assert m_c == expected_m_c
@@ -33,7 +52,9 @@ defmodule ExICE.CandidateTest do
     ip = {192, 168, 1, 1}
     port = 12_345
     m_c = "936255739 1 UDP 1234 192.168.1.1 12345 typ host"
-    expected_c = Candidate.new(:host, ip, port, nil, nil, nil, priority: 1234)
+
+    expected_c =
+      Candidate.new(:host, address: ip, port: port, priority: 1234, transport_module: nil)
 
     assert {:ok, c} = Candidate.unmarshal(m_c)
     c = %Candidate{c | id: expected_c.id}
